Implement MFT parsing. Implement Extractors
Refractor Code and implementation of Extractors. (json, csv, xml, neo4j)

MFT parsing works but will take a long time to process. The plaso files
containing MFTs info are 10+ Go.
Moreover using XML,JSON or CSV extractor, output files will be 20+Go.
(We parse our own meta and keep plaso's ones so it almost doubles the
size of the extract)

So in the end, MFT are implemented but it is not very useable